Adding support for 5dof (positional tracking) for DirectSBS mode.
A lot of work is still needed; but this commit adds support for positional
tracking by co-opting the cockpit/X/Y/Z/Reference fields in PlayerData table
and using the ship's heading as a coordinate system. It seems to work fine.

Support for TrackIR and SteamVR coming soon. I also need to migrate a bunch
of parameters from ddraw to this hook.