Added lean/look around with keyboard keys.
The Caps Lock toggles between leaning with the keyboard and looking
around. Look around with keys is only enabled when the mouse look is
disabled.

Also added global definitions to read the state of the Alt, Ctrl and
Shift keys and a new function to refresh these variables -- thanks to
Jeremy.