Release 1.0.4. Fixes to TrackIR when installed; but turned off.
In this case, TrackIR would write 0,0 to yaw,pitch because now I was setting
dataReady = true to enable cockpit inertia. I had to add one more flag to
prevent writing to yaw,pitch if TrackIR is off, so that we can still have
cockpit inertia; but still allow POV hat to look around.