Craft Inertia is now compatible with mouse look.
Mouse look can be used to move around and craft inertia is still enabled. The
trick is to save the last unmodified yaw/pitch and then apply inertia to it
before exiting the hook. When the hook gets called, the old pre-inertia values
are restored and saved again before applying inertia.

Also added code to limit the amount of inertia by normalizing yaw/pitch and then
clamping the length of the vector.