Fixed [*] key in external view when inertia is enabled.
The view can be modified using the *joystick* after pressing [*] in external
view; but if inertia is enabled, we ignore that, and reset the camera.

To avoid this, I'm computing the difference between the camera yaw/pitch at the
end of the hook, and the yaw/pitch the next time the hook is called. If there's
a difference, then the yaw/pitch was modified between calls to this hook, and
the pre-inertia yaw/pitch needs to be updated.