Fix the spinning-out-of-control external camera issue.
If we go to external view, so that the tilt is applied, then go back into the
cockpit and press '.' to remove the cockpit, and then we go back to exterior
view, the camera will spin out of control.

The problem was that if the cockpit is not displayed, then we don't save the
state of the yaw/pitch, so the logic will do weird things. The fix is to add
a case when the cockpit isn't displayed to save the yaw/pitch state. That way
we don't spin out of control.