Disable the keypad view if no joysticks are connected.
This change was requested by Jeremy because it apparently caused some conflicts
with this joystick hook. I added a conditional setting to disable this behavior
and revert to the old one in case it's needed. The new setting is:

test_joystick = 1 (Default) Disable keypad keys if no joysticks are present.
test_joystick = 0 Enable keypad keys even if no joysticks are connected.