Removed hard-coded paths in the project to use LocalProperties.props instead.
This should make the project easier to compile in other places. Machine-specific
paths should now be specified in the LocalProperties.props file.

In order to use the LocalProperties, each setting in the project must be set to
"Inherit from parent project".

I've put my properties in the LocalProperties.props.default file. To use it,
copy it to LocalProperties.props and load it in the project. Menu bar -> View ->
Other Windows -> Property Manager.