Adding a SharedMem object to share data with ddraw.
I'm using a standard implementation of a memory-mapped file in the Local
namespace. The shared data consists of a bool that tells us when the data is
valid and a void* that can be cast to anything we want and points to the heap.
Since ddraw and the hooks share the heap, we only need to pass this pointer
once.