Fix positional tracking jitter with new hook
Instead of modifying the camera shake values, I implemented a new hook to use the values for CockpitPosition before they are transformed from Object coordinates to World coordinates (CockpitPositionTransformed).

These values are float, so there is no loss of precision (float->int) that was the root cause of the jitter when using Shake.

Also there is no need to calculate or apply the heading matrix before injecting the offset into XWA.