Add fake tracking through the keyboard in SteamVR mode.
Positional tracking is now shared with ddraw too. This will help fix the
position of the HUD.

Rotational and positional tracking can now be faked through the keyboard in
SteamVR mode. This can be used along with the null SteamVR driver when debugging
VR. To enable this, add:

	keyboard_lean = 1
	keyboard_look = 1

to CockpitLook.cfg.