Fix tracking while in the hangar if pose corrected heatracking is ON.
The new hooks that get activated when pose corrected headtracking is ON do not
have any tracking information when we're in the hangar. That's why tracking is
broken. To patch this, I'm disabling the new hooks if we're in the hangar. This
"fix" needs corresponding code in ddraw to work properly.