Fix tracking for TrackIR by restoring old mouse look behavior.
Adding full 6dof tracking for TrackIR is possible, but it breaks a number of
things in ddraw. If we restore the old mouse look behavior, we have 6dof in
flight and 6dof in the hangar, except for the gunner turret. I think this is a
nice compromise without having to break ddraw again.

Notes:

If we decide to enable full 6dof for TrackIR later, the reticle will break. We
need to enable the ExternalHUDShader to fix the reticle. But that implies
capturing the reticle for TrackIR (it's currently only captured for VR modes).
The reticle's aspect ratio and scale will be wrong. The scale can be set to  0.5
and the aspect ratio can be fixed by allowing the code to capture the aspect
ratio in ComputeHyperFOVParams() for TrackIR as well.

We would also need to create/clear the reticle buffers (this is currently only
done if VR is enabled).

... and who knows what else! That's why it's just easier to re-enable mouse look
for TrackIR :)