Don't apply tilt to exterior views when parked in the hangar.
In external view, we apply tilt so that we can see the reticle and the ship at
the same time. However, doing so in the hangar messes up the camera. This is
particularly obvious when pressing '7' on the keyboard. The repro is:

	* Go to the Pilot Proving Grounds.
	* Select the X-Wing
	* Press 7, watch the X-Wing in external view off-center.
	* Go back to the cockpit, press 7 again and watch the X-Wing in the center
	  of the screen now.

This avoids applying the tilt in external view if we're parked in the hangar to
prevent this bug.