Add basic support for YawVR
Basic support for the motion simulation platform YawVR. To enable it, add the
following to CockpitLook.cfg:

	yawvr_enable   = 1
	yawvr_server   = 192.168.50.6
	yawvr_udp_port = 50010
	yawvr_tcp_port = 50020

	; Optional settings:
	yawvr_yaw_scale   = 7
	yawvr_roll_scale  = 3
	yawvr_pitch_scale = -200

The optional settings are values that scale the ypr inertia to sensible ypr
values for YawVR. Yaw is accumulated, so the seat can rotate a full 360 degrees;
but pitch and roll are not. So these values "spring back" when inertia is no
longer applied. Fun!