Add thread to re-center YawVR when the hook is not called. Remove hyper-jolt.
When the 2D interface is presented, this hook is no longer called, causing the
seat to be "stuck" in whatever position it was when flying. To fix this, I added
a thread that checks for a timestamp for the last update. If the timestamp is
longer than a few seconds, it will go back to its center position.

After exiting hyperspace, the sudden change in orientation is picked up by the
system and a big change in inertia is rendered. This translates to a
"hyper-jolt" which is probably bad for the YawVR simulator. To avoid unnecessary
wear-and-tear, I added a frame counter that gets reset when exiting hyperspace.
This counter will inhibit YawVR for 25 frames after exiting hyperspace.