Fix the external view after the map has been displayed.
The distance and orientation when the map is displayed will be applied to the
external view after the map is turned off. This should not be happening. This is
happening because External Inertia is applied while the map is displayed. Here
the fix is to simply avoid applying External Inertia while the map is displayed.

I have also added a new setting to CockpitLook.cfg that will disable this fix
and revert to the old behavior because some people are actually using the map
to set up shots of big starships/stations. To disable this fix, just add:

	enable_external_inertia_in_map = 1

To CockpitLook.cfg. This flag is disabled by default.