Add "shake" to the telemetry. This helps us detect all types of impacts.
Cockpit shake is activated in the following situations:

- Laser hits.
- Hull damage.
- Collisions against capital ships and other big objects.

It's not activated for collisions that cause the ship to roll; but hey, it's
something, right? This field can be used to drive bass shakers, but the range
must be clamped to prevent damage.

I don't know what's the maximum range for this field, but I've observed a
maximum value of around 80 units so far. More testing is needed here.